Rename back files to old naming and proper casing, after sucessfully detected change. Recover testMatch jest eslint rootDir config.
