+++ jest.eslint.config.js
@@ -13,6 +13,7 @@ module.exports = {
     'public/',
     'generated',
   ],
+  testMatch: ['<rootDir>/**/*.js', '<rootDir>/**/*.ts', '<rootDir>/**/*.tsx'],
   watchPlugins: [
     'jest-watch-typeahead/filename',
     'jest-runner-eslint/watch-fix',
