+++ src/components/carts/index.ts
@@ -1,7 +1,5 @@
 import { lazy } from 'react';
 
-const Carts = lazy(
-  () => import('./CartsScreen' /* webpackChunkName: "carts" */)
-);
+const Carts = lazy(() => import('./Carts' /* webpackChunkName: "carts" */));
 
 export default Carts;
